Moved template tags to wagtailspace.
