+++ wagtailio/project_styleguide/templates/patterns/includes/spaceheader.html
@@ -1,4 +1,4 @@
-{% load navigation_tags wagtailcore_tags static %}
+{% load wagtailspace_tags wagtailcore_tags static %}
 
 <header class="app__header header">
     {% include "patterns/components/skip-link/skip-link.html" %}
